+++ spec/spec_helper.rb
@@ -1,5 +1,11 @@
 # frozen_string_literal: true
 
+require "simplecov"
+SimpleCov.start do
+  add_filter "/spec/"
+  add_filter "/bin/"
+end
+
 require "e2e"
 require "e2e/rspec"